+++ components/engine/include/point.hpp
@@ -14,18 +14,18 @@
 
 class Point3D {
 public:
-  double x, y, z, w; // 3D coords of point
+  float x, y, z, w; // 3D coords of point
   Point3D() {
     x = y = z = 0;
     w = 1;
   }
-  Point3D(const double _x, const double _y, const double _z) {
+  Point3D(const float _x, const float _y, const float _z) {
     x = _x;
     y = _y;
     z = _z;
     w = 1;
   }
-  Point3D(const double _x, const double _y, const double _z, const double _w) {
+  Point3D(const float _x, const float _y, const float _z, const float _w) {
     x = _x;
     y = _y;
     z = _z;
@@ -41,11 +41,11 @@ public:
 
   Point3D &operator=(const Point3D &rhs) = default;
 
-  Point3D operator*(const double rhs) const;
+  Point3D operator*(const float rhs) const;
 
-  Point3D operator/(const double rhs) const;
+  Point3D operator/(const float rhs) const;
 
-  double operator*(const Point3D &rhs) const;
+  float operator*(const Point3D &rhs) const;
 
   Point3D operator+(const Point3D &rhs) const;
 
@@ -61,13 +61,13 @@ inline static Vector3D Cross(const Vector3D &a, const Vector3D &b) {
 }
 
 inline static Vector3D normalize(const Vector3D &rhs) {
-  double mag = sqrt(rhs.x * rhs.x + rhs.y * rhs.y + rhs.z * rhs.z);
+  float mag = sqrtf(rhs.x * rhs.x + rhs.y * rhs.y + rhs.z * rhs.z);
 
   return Vector3D(rhs.x / mag, rhs.y / mag, rhs.z / mag);
 }
 
-inline static double magnitude(const Vector3D &rhs) {
-  return sqrt(rhs.x * rhs.x + rhs.y * rhs.y + rhs.z * rhs.z);
+inline static float magnitude(const Vector3D &rhs) {
+  return sqrtf(rhs.x * rhs.x + rhs.y * rhs.y + rhs.z * rhs.z);
 }
 
 inline static Vector3D crossProduct(const Vector3D &a, const Vector3D &b) {
@@ -76,18 +76,18 @@ inline static Vector3D crossProduct(const Vector3D &a, const Vector3D &b) {
 
 class Point2D {
 public:
-  double x, y, w; // 2D coords of point
+  float x, y, w; // 2D coords of point
   Point2D() {
     x = y = 0;
     w = 1;
   }
   Point2D(const Point2D &rhs) = default;
-  Point2D(const double _x, const double _y) {
+  Point2D(const float _x, const float _y) {
     x = _x;
     y = _y;
     w = 1;
   }
-  Point2D(const double _x, const double _y, const double _w) {
+  Point2D(const float _x, const float _y, const float _w) {
     x = _x;
     y = _y;
     w = _w;
@@ -100,11 +100,11 @@ public:
 
   Point2D &operator=(const Point2D &rhs) = default;
 
-  Point2D operator*(const double rhs) const;
+  Point2D operator*(const float rhs) const;
 
-  Point2D operator/(const double rhs) const;
+  Point2D operator/(const float rhs) const;
 
-  double operator*(const Point2D &rhs) const;
+  float operator*(const Point2D &rhs) const;
 
   Point2D operator+(const Point2D &rhs) const;
 
@@ -113,4 +113,4 @@ public:
 
 typedef Point2D Vector2D;
 
-inline static double magnitude(const Vector2D &rhs) { return sqrt(rhs.x * rhs.x + rhs.y * rhs.y); }
+inline static float magnitude(const Vector2D &rhs) { return sqrtf(rhs.x * rhs.x + rhs.y * rhs.y); }
